+++ packages/kitchensink/service-fetchq/index.js
@@ -3,10 +3,18 @@ const serviceFastify = require('@forrestjs/service-fastify');
 const serviceFetchq = require('@forrestjs/service-fetchq');
 
 const featureHome = require('./feature-home');
+const featureQ1 = require('./feature-q1');
 
 runHookApp({
   trace: 'compact',
-  settings: {},
+  settings: {
+    fetchq: {
+      logLevel: 'error',
+      pool: {
+        max: 1,
+      },
+    },
+  },
   services: [serviceFetchq, serviceFastify],
-  features: [featureHome],
+  features: [featureHome, featureQ1],
 }).catch((err) => console.error(err));